requirements.txt
Flask
WTForms==1.0.5
Flask-admin==1.0.7
Flask-login
Flask-sqlalchemy
Flask-bcrypt
Flask-DebugToolbar
Flask-Script
Flask-OAuthlib
